Some new test.

+++ mpcat/mpcctl/systems/aircraft/install_aircraftpce/aircraftpce/include/arithmetic.h
@@ -2,7 +2,7 @@
 #define ARITHMETIC_H
 #ifndef __real_t_defined
 #define __real_t_defined
-typedef float32_t real_t;
+typedef float64_t real_t;
 #endif
 
 #endif
